ready for client

+++ server/src/res_times.rs
@@ -1,5 +1,5 @@
 pub struct TimeRes {
-    pub tick: u32,
+    pub tick: u64,
 }
 impl TimeRes {
     pub fn new() -> Self {
